implement test_join
